Refactor `BatchJobTable` component to follow SOLID principles by splitting it into smaller, manageable subcomponents.
